add flexibility to screen res of KC_SCREEN

+++ components/smsplus-esp32/smsplus-main.h
@@ -9,5 +9,5 @@
 
 void smsemuStart();
 
-uint32_t *vidGetOverlayBuf();
+uint16_t *vidGetOverlayBuf();
 void vidRenderOverlay();
